% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{dynamic_count_table_for_each_order}
\alias{dynamic_count_table_for_each_order}
\title{Build Dynamic Count Table}
\usage{
dynamic_count_table_for_each_order(x, N, alphabet)
}
\arguments{
\item{x}{Character vector of symbols.}

\item{N}{Maximum order.}

\item{alphabet}{Character vector of full alphabet.}
}
\value{
A list of length N+1.
Each element is a data.table for each order with columns:
index, context_id, Event, Ce, C, t, t1
}
\description{
Computes dynamic PPM counts for all timesteps and orders.
Counts are cumulative up to (but excluding) the current timestep.
}
