% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPM_R.R
\name{dynamicModel}
\alias{dynamicModel}
\title{Build dynamic count matrix}
\usage{
dynamicModel(x, N = 5, prior = NULL, ...)
}
\arguments{
\item{x}{Vector of tokens.}

\item{N}{Maximum N-gram length.}

\item{prior}{Optional prior model (countMatrix) to add counts to.}

\item{...}{Additional vectors of same length as \code{x}.}
}
\value{
A \code{data.table} with counts for each N-gram order.
}
\description{
Build dynamic count matrix
}
