% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{ppm_compute}
\alias{ppm_compute}
\title{PPM Wrapper Function}
\usage{
ppm_compute(
  x,
  alphabet,
  N = 3,
  ppm_type = c("backoff", "interpolation"),
  escape_func = escape_C,
  base_type = c("uniform", "unseen")
)
}
\arguments{
\item{x}{Character vector of events}

\item{alphabet}{Character vector of full alphabet}

\item{N}{Maximum order}

\item{ppm_type}{"backoff" or "interpolation"}

\item{escape_func}{Escape function (e.g., escape_C)}

\item{base_type}{"uniform" (1 / (∣alphabet∣) or "unseen" (1 / (∣alphabet∣+1−tseen), Harrison's style)}
}
\value{
data.table with columns: index, Event, P, IC, Entropy
}
\description{
Compute PPM probabilities with selectable backoff/interpolation and base probabilities.
}
