% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{precompute_contexts}
\alias{precompute_contexts}
\title{Precompute context identifiers for all orders}
\usage{
precompute_contexts(dt_lag, N)
}
\arguments{
\item{dt_lag}{Lag matrix from lag_matrix()}

\item{N}{Maximum n-gram order}
}
\value{
List of length N+1, each element is a character vector of context IDs
}
\description{
Precompute context identifiers for all orders
}
