% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backoff.R
\name{compute_local_probs}
\alias{compute_local_probs}
\title{Compute Local Probabilities for a Single Order; Used by Backoff PPM}
\usage{
compute_local_probs(dt, escape_func, normalize = FALSE)
}
\arguments{
\item{dt}{Data.table for a single order. Columns: index, context_id, Event, Ce, C, t, t1.}

\item{escape_func}{Escape function from escape.R.
Signature: `function(t, t1)` returning `list(subtract, esc_numer)`.}
}
\value{
Data.table with columns: index, Event, prob_local, esc
}
\description{
Compute Local Probabilities for a Single Order; Used by Backoff PPM
}
