% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{lag_matrix}
\alias{lag_matrix}
\title{Generate Lagged N-gram Matrix}
\usage{
lag_matrix(x, N = 3)
}
\arguments{
\item{x}{Character vector of symbols/events.}

\item{N}{Maximum N-gram order.}
}
\value{
A `data.table` with columns LagN..Lag0 and index.
}
\description{
Creates a lagged representation of a sequence for N-gram modeling.
}
